Making some distinction between PCI and PCIe
